laureano initial branchout

+++ resources/views/building/index.blade.php
@@ -30,11 +30,27 @@
                             <td>{{ $building->name }}</td>
                             <td>{{ $building->address }}</td>
                             <td>
-                                {!! Form::open(['url' => '/building/'.$building->id, 'method' => 'delete']) !!}
+                                <!-- Right Side Of Navbar -->
+                                <ul class="nav navbar-nav navbar-left">
+
+                                        <li class="dropdown">
+                                            <a href="#" class="dropdown-toggle" data-toggle="dropdown" role="button" aria-expanded="false" aria-haspopup="true">
+                                                options <span class="caret"></span>
+                                            </a>
+
+                                            <ul class="dropdown-menu">
+                                              <li>
+                                                  <a href="{{ route('building.edit', $building) }}" class="btn btn-primary">Modifier</a>
+                                              </li>
+                                              <li>
+                                                  <a href="{{ route('building.floor.index', $building) }}" class="btn btn-success">Étages</a>
+                                              </li>
+                                            </ul>
+                                        </li>
+                                </ul>
+                                {!! Form::open(['url' => '/building/'.$building->id,'onsubmit' => 'return ConfirmDelete()', 'method' => 'delete']) !!}
                                 {!! Form::submit('Supprimer', ['class' => 'btn btn-danger pull-right']) !!}
                                 {!! Form::close() !!}
-                                <a href="{{ route('building.edit', $building) }}" class="btn btn-primary">Modifier</a>
-                                <a href="{{ route('building.floor.index', $building) }}" class="btn btn-success">Étages</a>
                             </td>
                         </tr>
                         @endforeach
